move codelens to the providers folder
